test_UI_Anatoli_Kanavalchyk/requirements.txt
pytest
playwright
allure-pytest
faker
